requirements.txt
click==6.7
Flask==0.12.2
gunicorn==19.7.1
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
numpy==1.14.0
Pillow==5.0.0
Werkzeug==0.14.1
matplotlib==2.1.0
